requirements.txt
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
urllib3>=2.1.0
certifi>=2023.11.17
beautifulsoup4>=4.10.0